--- a/plugins/inputs/snmp/snmp_test.go
+++ b/plugins/inputs/snmp/snmp_test.go
@@ -1295,3 +1295,12 @@ func TestTableJoinNoIndexAsTag_walk(t *testing.T) {
 	require.Contains(t, tb.Rows, rtr2)
 	require.Contains(t, tb.Rows, rtr3)
 }
+
+func BenchmarkMibLoading(b *testing.B) {
+	log := testutil.Logger{}
+	path := []string{"testdata"}
+	for i := 0; i < b.N; i++ {
+		err := snmp.LoadMibsFromPath(path, log)
+		require.NoError(b, err)
+	}
+}