@Hipska 2021-12-02
### Current release
```log
2021-12-02T14:12:40Z I! Starting Telegraf 1.20.4
2021-12-02T14:12:41Z D! [agent] Initializing plugins
2021-12-02T14:12:41Z D! [agent] Connecting outputs
```
_Note_: This version still used `snmptranslate`  and `snmptable` on the fly when the input is started. (this takes around 10s on a fast system)

### Current master (1.21.0 RC0)
```log
2021-12-02T14:19:43Z I! Starting Telegraf 1.21.0
2021-12-02T14:19:44Z D! [agent] Initializing plugins
```
After this, I waited for 30 minutes until I killed it. CPU was high and memory usage grew steadily to 14GiB during that time.

### Artefact 0e707ebc
```log
2021-12-02T15:05:14Z I! Starting Telegraf 1.21.0-0e707ebc
2021-12-02T15:05:16Z D! [agent] Initializing plugins
```
Similar behaviour as RC0, stopped after 23 minutes and 10GiB of memory usage.