@srebhan 2021-12-02
With the added benchmark, i.e. loading the `testdata` directory multiple times I see the following

#### Current master
```text
goos: linux
goarch: amd64
pkg: github.com/influxdata/telegraf/plugins/inputs/snmp
cpu: AMD Ryzen 7 3700X 8-Core Processor             
BenchmarkMibLoading
BenchmarkMibLoading-16                 3         438131129 ns/op
PASS
```

#### Optimized locking
```text
goos: linux
goarch: amd64
pkg: github.com/influxdata/telegraf/plugins/inputs/snmp
cpu: AMD Ryzen 7 3700X 8-Core Processor             
BenchmarkMibLoading
BenchmarkMibLoading-16                 3         452756498 ns/op
PASS
```

#### Caching
```text
goos: linux
goarch: amd64
pkg: github.com/influxdata/telegraf/plugins/inputs/snmp
cpu: AMD Ryzen 7 3700X 8-Core Processor             
BenchmarkMibLoading
BenchmarkMibLoading-16           4237710               288.9 ns/op
PASS
```