@Hipska 2021-12-03
### Artefact 336b0ced
```log
2021-12-03T08:35:15Z I! Starting Telegraf 1.21.0-336b0ced
2021-12-03T08:35:16Z D! [agent] Initializing plugins
2021-12-03T08:35:21Z D! [agent] Connecting outputs
...
```
Whoa, this is what it should have been!